backend/src/requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-filter==2.4.0
django-import-export==2.5.0
django-multiselectfield==0.1.12
djangorestframework==3.12.2
et-xmlfile==1.0.1
idna==2.10
isort==5.6.4
jdcal==1.4.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.5
pep8==1.7.1
Pillow==8.0.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pylint==2.6.0
python-decouple==3.3
pytz==2020.5
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
xlrd==2.0.1
xlwt==1.3.0
